+++ Sources/SmithyEventStreams/Deserializers/SDKUnknownDeserializer.swift
@@ -51,11 +51,11 @@ struct SDKUnknownDeserializer: ShapeDeserializer {
         throw notImplemented
     }
 
-    func readInteger(_ schema: Schema) throws -> Int {
+    func readInteger(_ schema: Schema) throws -> Int32 {
         throw notImplemented
     }
 
-    func readLong(_ schema: Schema) throws -> Int {
+    func readLong(_ schema: Schema) throws -> Int64 {
         throw notImplemented
     }
